refactor: move encryption in helper folder
