+++ backend/utils/helper/encryption.js
